build(client-presence): split most type for coverage accuracy  - breaking change: `SessionClientStatus` constant (enum) renamed to `SessionClientStatusEnum`
Note that CJS build is left as single compile for simplicity and it is not used for code coverage.

As all externals are currently declarations there is no need for separation from the types compile. It does appear that this separation leads to eslint import/order rule making more sensible choice for import orderings.
