+++ packages/framework/presence/.eslintrc.cjs
@@ -7,7 +7,9 @@ module.exports = {
 	extends: [require.resolve("@fluidframework/eslint-config-fluid/strict"), "prettier"],
 	parserOptions: {
 		project: [
-			"./tsconfig.json",
+			"./tsconfig.externals.json",
+			"./tsconfig.runtime.json",
+			"./tsconfig.types.json",
 			"./src/test/tsconfig.json",
 			"./src/test/core-interfaces/tsconfig.no-exactOptionalPropertyTypes.json",
 		],
