+++ packages/framework/presence/api-extractor.json
@@ -1,6 +1,10 @@
 {
 	"$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
 	"extends": "../../../common/build/build-common/api-extractor-base.esm.no-legacy.json",
+	"projectFolder": "./",
+	"compiler": {
+		"tsconfigFilePath": "./tsconfig.runtime.json"
+	},
 	"messages": {
 		"extractorMessageReporting": {
 			// api-extractor does not appear able to resolve self references such as
