+++ packages/framework/presence/api-extractor/api-extractor-lint-alpha.esm.json
@@ -2,6 +2,10 @@
 	"$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
 	"extends": "<projectFolder>/../../../common/build/build-common/api-extractor-lint.entrypoint.json",
 	"mainEntryPointFilePath": "<projectFolder>/lib/alpha.d.ts",
+	"projectFolder": "../",
+	"compiler": {
+		"tsconfigFilePath": "../tsconfig.runtime.json"
+	},
 	"messages": {
 		"extractorMessageReporting": {
 			"ae-wrong-input-file-type": {
