+++ packages/framework/presence/api-extractor/api-extractor-lint-bundle.json
@@ -2,6 +2,10 @@
 	"$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
 	"extends": "<projectFolder>/../../../common/build/build-common/api-extractor-lint.json",
 	"mainEntryPointFilePath": "<projectFolder>/lib/index.d.ts",
+	"projectFolder": "../",
+	"compiler": {
+		"tsconfigFilePath": "../tsconfig.runtime.json"
+	},
 	"messages": {
 		"extractorMessageReporting": {
 			// api-extractor does not appear able to resolve self references such as
