+++ packages/framework/presence/src/latestMapValueManager.ts
@@ -5,25 +5,28 @@
 
 import { createEmitter } from "@fluid-internal/client-utils";
 import type { Listenable } from "@fluidframework/core-interfaces";
-
-import type { BroadcastControls, BroadcastControlSettings } from "./broadcastControls.js";
-import { OptionalBroadcastControl } from "./broadcastControls.js";
-import type { ValueManager } from "./internalTypes.js";
-import type {
-	LatestValueClientData,
-	LatestValueData,
-	LatestValueMetadata,
-} from "./latestValueTypes.js";
-import type { ClientSessionId, ISessionClient, SpecificSessionClient } from "./presence.js";
-import { datastoreFromHandle, type StateDatastore } from "./stateDatastore.js";
-import { brandIVM } from "./valueManager.js";
-
 import type {
 	JsonDeserialized,
 	JsonSerializable,
 } from "@fluidframework/presence/internal/core-interfaces";
 import type { InternalTypes } from "@fluidframework/presence/internal/exposedInternalTypes";
 import type { InternalUtilityTypes } from "@fluidframework/presence/internal/exposedUtilityTypes";
+
+import { OptionalBroadcastControl } from "./broadcastControls.js";
+import type { BroadcastControls, BroadcastControlSettings } from "./broadcastControlsTypes.js";
+import type { ValueManager } from "./internalTypes.js";
+import type {
+	LatestValueClientData,
+	LatestValueData,
+	LatestValueMetadata,
+} from "./latestValueTypes.js";
+import type {
+	ClientSessionId,
+	ISessionClient,
+	SpecificSessionClient,
+} from "./sessionClientTypes.js";
+import { datastoreFromHandle, type StateDatastore } from "./stateDatastore.js";
+import { brandIVM } from "./valueManager.js";
 
 /**
  * Collection of latest known values for a specific client.
