+++ packages/framework/presence/src/latestValueManager.ts
@@ -5,22 +5,21 @@
 
 import { createEmitter } from "@fluid-internal/client-utils";
 import type { Listenable } from "@fluidframework/core-interfaces";
-
-import type { BroadcastControls, BroadcastControlSettings } from "./broadcastControls.js";
-import { OptionalBroadcastControl } from "./broadcastControls.js";
-import type { ValueManager } from "./internalTypes.js";
-import { brandedObjectEntries } from "./internalTypes.js";
-import type { LatestValueClientData, LatestValueData } from "./latestValueTypes.js";
-import type { ISessionClient } from "./presence.js";
-import { datastoreFromHandle, type StateDatastore } from "./stateDatastore.js";
-import { brandIVM } from "./valueManager.js";
-
 import type {
 	JsonDeserialized,
 	JsonSerializable,
 } from "@fluidframework/presence/internal/core-interfaces";
 import type { InternalTypes } from "@fluidframework/presence/internal/exposedInternalTypes";
 import type { InternalUtilityTypes } from "@fluidframework/presence/internal/exposedUtilityTypes";
+
+import { OptionalBroadcastControl } from "./broadcastControls.js";
+import type { BroadcastControls, BroadcastControlSettings } from "./broadcastControlsTypes.js";
+import type { ValueManager } from "./internalTypes.js";
+import { brandedObjectEntries } from "./internalUtils.js";
+import type { LatestValueClientData, LatestValueData } from "./latestValueTypes.js";
+import type { ISessionClient } from "./sessionClientTypes.js";
+import { datastoreFromHandle, type StateDatastore } from "./stateDatastore.js";
+import { brandIVM } from "./valueManager.js";
 
 /**
  * @sealed
