+++ packages/framework/presence/src/latestValueTypes.ts
@@ -3,10 +3,10 @@
  * Licensed under the MIT License.
  */
 
-import type { ISessionClient } from "./presence.js";
-
 import type { JsonDeserialized } from "@fluidframework/presence/internal/core-interfaces";
 import type { InternalUtilityTypes } from "@fluidframework/presence/internal/exposedUtilityTypes";
+
+import type { ISessionClient } from "./sessionClientTypes.js";
 
 /**
  * Metadata for the value state.
