+++ packages/framework/presence/src/stateDatastore.ts
@@ -6,7 +6,7 @@
 import type { ClientConnectionId } from "./baseTypes.js";
 import type { InternalTypes } from "./exposedInternalTypes.js";
 import type { ClientRecord } from "./internalTypes.js";
-import type { ClientSessionId, ISessionClient } from "./presence.js";
+import type { ClientSessionId, ISessionClient } from "./sessionClientTypes.js";
 
 // type StateDatastoreSchemaNode<
 // 	TValue extends InternalTypes.ValueDirectoryOrState<any> = InternalTypes.ValueDirectoryOrState<unknown>,
