+++ packages/framework/presence/src/test/core-interfaces/tsconfig.no-exactOptionalPropertyTypes.json
@@ -12,7 +12,7 @@
 	},
 	"references": [
 		{
-			"path": "../../..",
+			"path": "../../../tsconfig.externals.json",
 		},
 	],
 }
