+++ packages/framework/presence/src/test/tsconfig.json
@@ -13,7 +13,7 @@
 	},
 	"references": [
 		{
-			"path": "../..",
+			"path": "../../tsconfig.runtime.json",
 		},
 	],
 }
