+++ packages/framework/presence/tsconfig.base.json
@@ -1,7 +1,5 @@
 {
 	"extends": "../../../common/build/build-common/tsconfig.node16.json",
-	"include": ["src/**/*"],
-	"exclude": ["src/test/**/*"],
 	"compilerOptions": {
 		"rootDir": "./src",
 		"outDir": "./lib",
